% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_time_to_event.R
\name{predict_time_to_event_model}
\alias{predict_time_to_event_model}
\title{Make time-to-event predictions}
\usage{
predict_time_to_event_model(
  stan_model,
  start_date = 1754,
  end_date = 2021,
  no_geoloc = NULL,
  outputdir = NULL,
  event = "description",
  nsamp = NULL
)
}
\arguments{
\item{stan_model}{A path to the file (.parquet) containing the time-to-event data for species in a given region with valid occurrence records.}

\item{start_date}{A numeric specifying the year of reference the time to species description follows from. Default is 1754. Only used if `event='description'`.}

\item{end_date}{A numeric specifying the year at which to make inference about the time to species description. Only used if `event='description'`.}

\item{no_geoloc}{A numeric specifying the number of species in the region (e.g. botanical country) without valid geographic coordinates. Only used if `event='geolocation'`.}

\item{outputdir}{A path to a directory to save the predictions. Default is parent of the current directory}

\item{event}{A string specifying the event to model. It must be either `description` or `geolocation`.}
}
\value{
An integer number. `0` in case of success otherwise `-1`
}
\description{
Utility function to make prediction from a time-to-event model
}
