% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_time_to_event.R
\name{run_time_to_event_model}
\alias{run_time_to_event_model}
\title{Run a time-to-event model}
\usage{
run_time_to_event_model(
  occ_data,
  gap_data,
  outputdir = NULL,
  event = "description",
  end_date = NULL,
  evaluate = TRUE,
  predict = TRUE
)
}
\arguments{
\item{occ_data}{A path to the file (.parquet) containing the time-to-event data for species in a given region with valid occurrence records.}

\item{gap_data}{A path to the file (.parquet) containing the time-to-event data for species in the same region as `occ_data`, but with no
valid occurrence record}

\item{outputdir}{A path to a directory to save modelling outputs. Default is current directory}

\item{event}{A string specifying the event to model. It must be either `description` or `geolocation`.}

\item{end_date}{A numeric specifying the year at which to make inference about the time to species description.}

\item{evaluate}{A logical. Should the model calibration be evaluated ? Default is `TRUE`}

\item{predict}{A logical. Should the model make (new) predictions ?  Default is `TRUE`.}
}
\value{
An integer number. `0` in case of success otherwise `-1`
}
\description{
Utility function to run a time-to-event model
}
